Rename stuff to be more conventional and clear
