+++ server/lib/backend-adapters/index.js
@@ -4,7 +4,7 @@ import FastFtFeed from './fast-ft';
 import CAPI from './capi';
 import MockCapi from './mock-capi';
 import Hui from './hui';
-import Popular from './popular';
+import PopularFTContent from './popular-ft-content';
 import Liveblog from './liveblog';
 import MockLiveblog from './mock-liveblog';
 import Video from './video';
@@ -23,7 +23,7 @@ const liveblog = new Liveblog(memCache);
 const mockLiveblog = new MockLiveblog(liveblog);
 const myft = new Myft(memCache);
 const popularApi = new PopularAPI(memCache);
-const popular = new Popular(memCache);
+const popularFTContent = new PopularFTContent(memCache);
 const video = new Video(memCache);
 
 export default (flags = {}) => ({
@@ -33,7 +33,7 @@ export default (flags = {}) => ({
 	liveblog: flags.mockFrontPage ? mockLiveblog : liveblog,
 	myft,
 	popularApi,
-	popular,
+	popularFTContent,
 	video
 });
