+++ server/lib/backend-adapters/popular-api.js
@@ -1,4 +1,4 @@
-import sliceList from '../helpers/sliceList';
+import sliceList from '../helpers/slice-list';
 
 class PopularAPI {
 	constructor(cache) {
