+++ server/lib/helpers/filter-content.js
@@ -1,6 +1,6 @@
-import sliceList from '../helpers/sliceList';
+import sliceList from '../helpers/slice-list';
 import articleGenres from 'ft-next-article-genre';
-import capifyMetadata from '../helpers/capifyMetadata';
+import capifyMetadata from '../helpers/capify-metadata';
 
 // internal content filtering logic shared for ContentV1 and ContentV2
 export default ({from, limit, genres, type}, resolveType) => {
