+++ server/lib/helpers/slice-list.js
