+++ server/lib/types/content.js
@@ -1,7 +1,7 @@
 import articleGenres from 'ft-next-article-genre';
 import articleBranding from 'ft-n-article-branding';
 
-import capifyMetadata from '../helpers/capifyMetadata';
+import capifyMetadata from '../helpers/capify-metadata';
 
 import backend from '../backend-adapters/index';
